Updated the code based on the review comments provided by @aldettinger
