--- aws-lambda/.dockerignore
@@ -1,5 +1,0 @@
-*
-!target/*-runner
-!target/*-runner.jar
-!target/lib/*
-!target/quarkus-app/*
