+++ aws-lambda/pom.xml
@@ -13,38 +13,35 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 -->
+
 <project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>io.quarkus.camel</groupId>
-  <artifactId>aws-lambda</artifactId>
-  <version>1.0.0-SNAPSHOT</version>
+  <groupId>org.apache.camel.quarkus.examples</groupId>
+  <artifactId>camel-quarkus-examples-aws-lambda</artifactId>
+  <version>2.2.0-SNAPSHOT</version>
+
+  <name>Camel Quarkus :: Examples :: AWS Lambda</name>
+  <description>Camel Quarkus Example :: Deploying a Camel Route in AWS Lambda</description>
+
   <properties>
-    <compiler-plugin.version>3.8.1</compiler-plugin.version>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+    <quarkus.platform.groupid>io.quarkus.platform</quarkus.platform.groupid>
+    <quarkus.platform.version>2.1.0.Final</quarkus.platform.version>
+    <camel.quarkus.version>2.1.0</camel.quarkus.version>
+    <maven.compiler.plugin.version>3.8.1</maven.compiler.plugin.version>
     <maven.compiler.parameters>true</maven.compiler.parameters>
     <maven.compiler.source>11</maven.compiler.source>
     <maven.compiler.target>11</maven.compiler.target>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    <quarkus.platform.artifact-id>quarkus-bom</quarkus.platform.artifact-id>
-    <quarkus.platform.group-id>io.quarkus.platform</quarkus.platform.group-id>
-    <quarkus.platform.version>2.1.0.Final</quarkus.platform.version>
-    <camel-quarkus.version>2.1.0</camel-quarkus.version>
-    <surefire-plugin.version>3.0.0-M5</surefire-plugin.version>
+    <maven.surefire.plugin.version>3.0.0-M5</maven.surefire.plugin.version>
   </properties>
   <dependencyManagement>
     <dependencies>
       <dependency>
-        <groupId>${quarkus.platform.group-id}</groupId>
-        <artifactId>${quarkus.platform.artifact-id}</artifactId>
-        <version>${quarkus.platform.version}</version>
-        <type>pom</type>
-        <scope>import</scope>
-      </dependency>
-      <dependency>
         <groupId>org.apache.camel.quarkus</groupId>
         <artifactId>camel-quarkus-bom</artifactId>
-        <version>${camel-quarkus.version}</version>
+        <version>${camel.quarkus.version}</version>
         <type>pom</type>
         <scope>import</scope>
       </dependency>
@@ -94,7 +91,7 @@
   <build>
     <plugins>
       <plugin>
-        <groupId>${quarkus.platform.group-id}</groupId>
+        <groupId>${quarkus.platform.groupid}</groupId>
         <artifactId>quarkus-maven-plugin</artifactId>
         <version>${quarkus.platform.version}</version>
         <extensions>true</extensions>
@@ -110,14 +107,14 @@
       </plugin>
       <plugin>
         <artifactId>maven-compiler-plugin</artifactId>
-        <version>${compiler-plugin.version}</version>
+        <version>${maven.compiler.plugin.version}</version>
         <configuration>
           <parameters>${maven.compiler.parameters}</parameters>
         </configuration>
       </plugin>
       <plugin>
         <artifactId>maven-surefire-plugin</artifactId>
-        <version>${surefire-plugin.version}</version>
+        <version>${maven.surefire.plugin.version}</version>
         <configuration>
           <systemPropertyVariables>
             <java.util.logging.manager>org.jboss.logmanager.LogManager</java.util.logging.manager>
@@ -139,7 +136,7 @@
         <plugins>
           <plugin>
             <artifactId>maven-failsafe-plugin</artifactId>
-            <version>${surefire-plugin.version}</version>
+            <version>${maven.surefire.plugin.version}</version>
             <executions>
               <execution>
                 <goals>
