+++ aws-lambda/src/main/docker/Dockerfile.native-distroless
@@ -3,7 +3,7 @@
 #
 # Before building the container image run:
 #
-# ./mvnw package -Pnative
+# mvn package -Pnative
 #
 # Then, build the image with:
 #
