+++ aws-lambda/src/main/java/org/apache/camel/quarkus/examples/CamelRoute.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.quarkus.camel;
+package org.apache.camel.quarkus.examples;
 
 import org.apache.camel.builder.RouteBuilder;
 
@@ -31,12 +31,6 @@ public class CamelRoute extends RouteBuilder {
     public void configure() throws Exception {
         from("direct:input").routeId("Test")
            .log("Inside Camel Route Received Payload ==> ${body}")
-           /*
-            * If you use the below Bean EIP then it uses reflection to invoke the bean method at runtime.
-            * Avoiding Reflection will have some improvement in terms of boot up, execution time and max memory used
-            *
-            */
-           //.bean(greetService,"greet(${body.name})")
            .setBody().body(Person.class, p -> greetService.greet(p.getName()))
         .end();
     }
