+++ aws-lambda/src/main/java/org/apache/camel/quarkus/examples/Person.java
@@ -14,26 +14,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.quarkus.camel;
+package org.apache.camel.quarkus.examples;
 
-import com.amazonaws.services.lambda.runtime.Context;
-import com.amazonaws.services.lambda.runtime.LambdaLogger;
-import com.amazonaws.services.lambda.runtime.RequestHandler;
-import org.apache.camel.ProducerTemplate;
+public class Person {
 
-import javax.inject.Inject;
-import javax.inject.Named;
+    private String name;
 
-@Named("awsLambdaHandler")
-public class AWSLambdaHandler implements RequestHandler<Person, String> {
+    public String getName() {
+        return name;
+    }
 
-    @Inject
-    ProducerTemplate template;
+    public Person setName(String name) {
+        this.name = name;
+        return this;
+    }
 
     @Override
-    public String handleRequest(Person input, Context context) {
-        LambdaLogger logger = context.getLogger();
-        logger.log("Calling Camel Route :)");
-        return template.requestBody("direct:input", input, String.class);
+    public String toString() {
+        return String.format("Person[name= %s]",this.name);
     }
 }
