+++ aws-lambda/src/test/java/org/apache/camel/quarkus/examples/AWSLambdaHandlerTestIT.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.quarkus.camel;
+package org.apache.camel.quarkus.examples;
 
 import io.quarkus.test.junit.NativeImageTest;
